backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pinecone-client==3.0.0
python-dotenv==1.0.0
# Face recognition dependencies (should be pre-installed in animcogn/face_recognition)
# face-recognition==1.3.0
# opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.4